--- a/caffe2/operators/mish_op.h
+++ b/caffe2/operators/mish_op.h
@@ -22,12 +22,12 @@ class MishGradientOp final : public Operator<Context> {
   bool DoRunWithType();
 
   bool RunOnDevice() override {
-    return DispatchHelper<TensorTypes<float, double>>::call(this, Input(X));
+    return DispatchHelper<TensorTypes<float, double>>::call(this, Input(INPUT));
   }
 
- protected:
-  INPUT_TAGS(X, Y, DY);
-  OUTPUT_TAGS(DX);
+ private:
+  INPUT_TAGS(INPUT, OUTPUT, OUTPUT_GRAD);
+  OUTPUT_TAGS(INPUT_GRAD);
 };
 
 } // namespace caffe2