--- a/caffe2/python/operator_test/activation_ops_test.py
+++ b/caffe2/python/operator_test/activation_ops_test.py
@@ -263,5 +263,32 @@ class TestActivations(serial.SerializedTestCase):
                                   ensure_outputs_are_inferred=True)
 
 
+    @given(n=st.integers(0, 6), m=st.integers(4, 6),
+           seed=st.integers(0, 1000), **hu.gcs_cpu_only)
+    def test_mish(self, n, m, gc, dc, seed):
+        np.random.seed(seed)
+        X = np.random.rand(n, m).astype(np.float32)
+
+        def mish_ref(X):
+            return (X * np.tanh(np.log1p(np.exp(X))),)
+
+        op = core.CreateOperator(
+            "Mish",
+            ["X"],
+            ["Y"]
+        )
+
+        self.assertReferenceChecks(
+            device_option=gc,
+            op=op,
+            inputs=[X],
+            reference=mish_ref,
+            ensure_outputs_are_inferred=True,
+        )
+
+        self.assertGradientChecks(
+            gc, op, [X], 0, [0], ensure_outputs_are_inferred=True)
+
+
 if __name__ == "__main__":
     unittest.main()