… caffe2/utils/math/elementwise.cc
… caffe2/utils/math/elementwise.cu