Summary: This diff improves Caffe2 MishOp as follows:
1. Support double type.
2. Improve forward performance on CPU.
3. Improve backward performance on CPU.
4. Using real backward formula instead of the approximate one.

Differential Revision: D25071304

Benchmark:
input size = [4096, 32768]
forward 929ms -> 626ms
backward 1711ms -> 844ms